requirements.txt
langchain
langchain_community
langchain-anthropic
anthropic
python-dotenv
requests